[refactor] Mirror /courses layout on /projects, fix Greptile feedback
- Extract CourseScheduleMenu/List/Card from /courses/index.tsx into a shared
  components/courses/CourseSchedule.tsx. /courses and /projects both consume
  it. The breadcrumb menu now takes a `sections` array so each page composes
  its own (Courses+Projects vs Projects-only).
- /projects now renders the same sticky breadcrumb sidebar + cohort/rounds
  cards that /courses uses, with an empty-state message when no projects are
  active.
- Remove the /projects → blog.bluedot.org redirect from next.config.js so
  the new page is reachable. The /projects/:slug → blog redirect stays for
  legacy inbound post links.
- Pass loading={programsLoading} to the Programs nav dropdown so it shows
  ProgressDots in flight, matching Courses + Projects (Greptile P2).
