+++ apps/website/next.config.js
@@ -19,11 +19,6 @@ module.exports = withDefaultBlueDotNextConfig({
         permanent: true,
       },
       {
-        source: '/projects',
-        destination: 'https://blog.bluedot.org/s/projects',
-        permanent: true,
-      },
-      {
         source: '/projects/:slug',
         destination: 'https://blog.bluedot.org/p/:slug',
         permanent: true,
